Consolidation of files, resolution of review comments
