+++ server/api/signup.js
@@ -8,23 +8,25 @@ const Crypto = require("crypto");
 /**
  * The callback function for the /api/signup POST route. Signs up a user into the database.
  *
- * The `res` parameter should contain both `username` and `password` in its body.
+ * The `req` parameter should contain both `username` and `password` in its body.
  * @param {Express.Request} req - The Request object generated by the route.
  * @param {Express.Response} res - The Result object generated by the route.
  * @param {Function} next - The next() function used in the callback.
  * @returns Nothing.
  */
 function signup(req, res, next) {
+  var { username, password } = req.body;
+
   // check that signup data is real
-  if (!req.body.username || !req.body.password) {
+  if (!username || !password) {
     res.status(400).send("Error: Username or password not set!");
     return;
   }
 
   // test username and password; should throw an error if there's an issue
   try {
-    checkUsernameRequirements(req.body.username);
-    checkPasswordRequirements(req.body.password);
+    checkUsernameRequirements(username);
+    checkPasswordRequirements(password);
   } catch (error) {
     res.status(401).send(error.message);
     return;
@@ -33,7 +35,7 @@ function signup(req, res, next) {
   const salt = Crypto.randomBytes(16); // salt is required for hashing; it makes
   // otherwise identical hashes different
   Crypto.pbkdf2(
-    req.body.password,
+    password,
     salt,
     310000,
     32,
@@ -48,13 +50,11 @@ function signup(req, res, next) {
       try {
         const newUser = await insertUser(
           db,
-          req.body.username,
+          username,
           hashedPassword,
           salt,
           (user) => {
-            res
-              .status(200)
-              .send(`Successfully signed user ${req.body.username} up!`);
+            res.status(200).send(`Successfully signed user ${username} up!`);
           }
         );
         // TODO (integration): After correctly signing up, log the user in
